Modified deployment template to use Lambda Layers
The content produced by this repository will be deployed as a shared Lambda Layer. Handlers then deploy with this Layer as a dependency.

The intent here is to allow centralised upgrades for the common wrapper code in this plugin package.

The Handlers.yaml template is now extended to include all CRUDL handlers as functions provided by a single package.