Introducing platformCredentials via Request
In order to allow the runtime to execute with a minimum privelege role we must pass through credentials from the calling platform (i.e; CloudFormation). This change means that the various AWS SDK Clients cannot be initialised as part of the function construction flow. This is an effect of how SDK clients are passed credentials as part of their construction (there is no way to post-inject credentials besides via ENV_VARS which is exactly what we are preventing here). Hence, I introduced a new `initialiseRuntime` function to create the various AWS clients and things which depend on them.

These changes also led me to reconsider the use of Guice - there was very little left which could be injected by default and so rather than persist with Guice for DI, I have removed it entirely from the plugin. Dependencies are created with defaults, and explicit constructor injection for testing purposes remains in place.

Lastly - all test content has been updated to include the new credentials block and additional tests added for both missing credentials and for the default injection path. I have also locally initialised a new Java handler package with RPDK and tested it and it seems to be functioning correctly.