Catches Throwable to handle unchecked exceptions
During testing, we experienced a case of an uncaught Throwable bubbling out as a NullPointerException in the response handling. This made it impossible to understand that actual root cause of a provider failure. Have changed the 'catch all' logic to account for Throwable rather than just the subclass Exception type.