Adding Helpers for help with Exception handling, retries and replay (#112)
* Adding helpers for retry, error handling and replay semantics
that all handlers can leverage

* Adding testing, sample service and handlers inside test

* Improved Test coverage to meet standards
Fixed bugs from testing

* Bug fixes

* Few more delay types

* Few more delay types
Improve consumer ? super T clauses

* Refactored and improved tests across the board (#94)

* Renamed test methods

* Refactored verify( ... , times(0)) statements

* Refactored response testing and standard errors

Modified the raw OutputStream test asserts to use modelled HandlerResponse checks. In doing so, found a few errors with the way responses were serialized.

Also identified some small improvements for handling and reporting of errors, such as ValidationExceptions and addressed those in code and test.

* Upgraded test dependencies for codegen (#98)

Updated test dependencies for the codegen'd provider packages to JUnit5 and AssertJ.

* Catches Throwable to handle unchecked exceptions (#105)

* Catches Throwable to handle unchecked exceptions

During testing, we experienced a case of an uncaught Throwable bubbling out as a NullPointerException in the response handling. This made it impossible to understand that actual root cause of a provider failure. Have changed the 'catch all' logic to account for Throwable rather than just the subclass Exception type.

MetricsPublisherImpl continues to catch Exception only, allowing Errors to propagate to the main handler logic. Added a test to verify FAILED response is emitted back to caller.

* Use default name for SAM template (#106)

* Improved Test coverage to meet standards
Fixed bugs from testing

* Fix pom.xml to 2.5.11

* Improved Test coverage to meet standards
Fixed bugs from testing

* Fix compile error from merge

* Fix spacing for the json file for passing linting

* Reverting incorrect changes

* Reverted another incorrect change

* Bumped coverage benchmark

* Bug Fix: Prevent request caching when the actual call to service fails.
This ensures that one can resume from failed state, by fixing the error
and continuing the operation. This allows for CloudFormation to resume from
failed state for a stack by replaying the context.

Changed StdCallbackContext to preseve insertion order for the call
graphs. This makes it easier to understand the call graph order in code
with the order that is stashed inside the call context.

* Addressed comments specified on PR.

* Migrated to new error codes