Improved validation for loose serialization
Changes validation logic to validate the JSON output which has been serialized through a ResourceModel, rather than validating the raw payload. This allows handlers to accept string-quoted payloads which are technically invalid as raw JSON, but can serialize correctly to a typed POJO.

Uses a strict serializer which does not ignore extraneous fields, for the model serialization and converts the serialization error into a ValidationException when a resource properties payload includes fields which are not defined in the schema.

Also adds a secret to .gitallowed which was omitted in some prior PR and was impacting my pre-commit