Add exception messages to HandlerErrorCodes
This adds the messages to the error codes directly rather than having
them in the exception classes. This is good to unify the messages in
case other exception classes reuse the same code, allows users to access
the message directly, and also makes it easier to generate exception
classes if we decide to.