Add support for stack and change set level hooks
* Add StackHookTargetModel for support of stack-level hooks

Add new type that represents the payload sent to hook handlers that will contain information related to stack hooks.

* Add support to download hook target data for stack-level hooks

Stack level hooks will not be provided with invocation payload information, unlike resource level hooks. Instead, Hooks Service will pass in an S3 presigned URL that points to a file that contains the stack-level invocation payload. The base hook handler (before it reaches the customer's handler code), will use that URL to download the data and set it on the target model that is passed to the customer's handler.

* Add support to download hook target data for stack-level hooks

Stack level hooks will not be provided with invocation payload information, unlike resource level hooks. Instead, Hooks Service will pass in an S3 presigned URL that points to a file that contains the stack-level invocation payload. The base hook handler (before it reaches the customer's handler code), will use that URL to download the data and set it on the target model that is passed to the customer's handler.

* Add support to download hook target data for stack-level hooks

Stack level hooks will not be provided with invocation payload information, unlike resource level hooks. Instead, Hooks Service will pass in an S3 presigned URL that points to a file that contains the stack-level invocation payload. The base hook handler (before it reaches the customer's handler code), will use that URL to download the data and set it on the target model that is passed to the customer's handler.

* Skip stack level hook for stack if prior stack level change set hook succeeded

For stack level hooks, customers are able to return a new status that allow stack level hooks that execute against a stack to skip with a successful status. The idea is that if a stack hook invoked against a change set succeeds, there is no need to invoke against the stack once the change set is processed.

* Skip stack level hook for stack if prior stack level change set hook succeeded

For stack level hooks, customers are able to return a new status that allow stack level hooks that execute against a stack to skip with a successful status. The idea is that if a stack hook invoked against a change set succeeds, there is no need to invoke against the stack once the change set is processed.

* fix method

* Fix resource targetting for a stack level hook

* Fix resource targetting for a stack level hook

* bump version